feat(auth): disable /forgot with 'not available yet' notice until mail transport
The /forgot page let users submit reset/magic-link requests that go nowhere a
user can reach until a mail transport is configured (#277) — the link is only
logged server-side. Gate the page on env.MAIL_WEBHOOK_URL: when no transport is
configured, show a 'password reset isn't available yet' notice and disable the
email field and both submit buttons (greyed). When configured, the page works
as before.

- Split the page into a server wrapper (reads env, force-dynamic so prod can
  enable delivery via env without a rebuild) and a client ForgotForm taking a
  mailEnabled prop.
- Add auth.forgot_unavailable (en/nl).
- Component test covers the disabled and enabled states.
- playwright webServer sets a stub MAIL_WEBHOOK_URL so the recovery UI e2e flow
  stays enabled (delivery fails silently, the route still returns ok).

Closes #282
